+++ src/lib/addon.ts
@@ -218,7 +218,7 @@ export async function buildCatalogHandler(
   try {
     const { id, type } = args;
     const { skip, search } = args.extra;
-    const catalogKey = `catalog:${type}:${id}:${skip}:${search}:${config.nsfw}`;
+    const catalogKey = `catalog:${type}:${id}:${skip}:${search}:${config.nsfw}:${config.poster}`;
     const cacheCatalog = cache.get(catalogKey);
     if (cacheCatalog) return cacheCatalog;
     const filteredProviders = filterProvider(providers, id, config, "catalog");